switch back to `void*` by using a patch to modify `imconfig.h`

+++ src/graphic/Fast3D/gfx_direct3d11.cpp
@@ -1229,8 +1229,8 @@ gfx_d3d11_get_pixel_depth(int fb_id, const std::set<std::pair<float, float>>& co
 
 } // namespace
 
-ImTextureID gfx_d3d11_get_texture_by_id(unsigned long long id) {
-    return (unsigned long long)d3d.textures[id].resource_view.Get();
+ImTextureID gfx_d3d11_get_texture_by_id(int id) {
+    return d3d.textures[id].resource_view.Get();
 }
 
 void gfx_d3d11_enable_srgb_mode() {
